backend/requirements.txt
requests
opencv-python
deepforest
matplotlib
python-dotenv